pyproject.toml
[project]
name = "talentpilot"
version = "0.1.0"
description = "Automated job application pipeline for LinkedIn Easy Apply"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "playwright>=1.40.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
start = "python -m talentpilot"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
